+++ extevents/extevents.go
@@ -15,12 +15,16 @@ import (
 	"github.com/steadybit/extension-kit/exthttp"
 	"github.com/steadybit/extension-kit/extutil"
 	"net/http"
+	"strconv"
+	"sync"
 	"time"
 )
 
 func RegisterEventListenerHandlers() {
 	exthttp.RegisterHttpHandler("/events/experiment-started", onExperimentStarted)
 	exthttp.RegisterHttpHandler("/events/experiment-completed", onExperimentCompleted)
+	exthttp.RegisterHttpHandler("/events/experiment-step-started", onExperimentStepStarted)
+	exthttp.RegisterHttpHandler("/events/experiment-step-completed", onExperimentStepCompleted)
 }
 
 type SendEventApi interface {
@@ -39,8 +43,9 @@ func onExperimentStarted(w http.ResponseWriter, r *http.Request, body []byte) {
 	}
 
 	datadogEventBody := datadogV1.EventCreateRequest{
-		Title: fmt.Sprintf("Experiment '%s' (execution ID %.0f) started", event.ExperimentExecution.Name, event.ExperimentExecution.ExecutionId),
-		Text: fmt.Sprintf("%%%%%% \nThe chaos engineering experiment `%s` (execution %.0f) started.\n\nThe experiment is executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event).\n %%%%%%",
+		Title: fmt.Sprintf("Experiment '%s' started", event.ExperimentExecution.ExperimentKey),
+		Text: fmt.Sprintf("%%%%%% \nExperiment `%s` - `%s` (execution `%.0f`) started.\n\n_The experiment is executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event)._\n %%%%%%",
+			event.ExperimentExecution.ExperimentKey,
 			event.ExperimentExecution.Name,
 			event.ExperimentExecution.ExecutionId),
 		Tags:           tags,
@@ -58,6 +63,14 @@ func onExperimentCompleted(w http.ResponseWriter, r *http.Request, body []byte) 
 		exthttp.WriteError(w, extension_kit.ToError("Failed to decode event request body", err))
 		return
 	}
+
+	lastStartedStepsMux.Lock()
+	delete(lastStartedSteps, event.ExperimentExecution.ExecutionId)
+	lastStartedStepsMux.Unlock()
+	lastCompletedStepsMux.Lock()
+	delete(lastCompletedSteps, event.ExperimentExecution.ExecutionId)
+	lastCompletedStepsMux.Unlock()
+
 	tags := convertSteadybitEventToDataDogEventTags(event)
 	if tags == nil {
 		return
@@ -65,8 +78,9 @@ func onExperimentCompleted(w http.ResponseWriter, r *http.Request, body []byte) 
 
 	duration := event.ExperimentExecution.EndedTime.Sub(event.ExperimentExecution.PreparedTime)
 	datadogEventBody := datadogV1.EventCreateRequest{
-		Title: fmt.Sprintf("Experiment '%s' (execution ID %.0f) ended", event.ExperimentExecution.Name, event.ExperimentExecution.ExecutionId),
-		Text: fmt.Sprintf("%%%%%% \nThe chaos engineering experiment `%s` (execution %.0f) ended with state `%s` after %.2f seconds.\n\nThe experiment was executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event).\n %%%%%%",
+		Title: fmt.Sprintf("Experiment '%s' ended", event.ExperimentExecution.ExperimentKey),
+		Text: fmt.Sprintf("%%%%%% \nExperiment `%s` - `%s` (execution `%.0f`) ended with state `%s` after %.2f seconds.\n\n_The experiment is executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event)._\n %%%%%%",
+			event.ExperimentExecution.ExperimentKey,
 			event.ExperimentExecution.Name,
 			event.ExperimentExecution.ExecutionId,
 			event.ExperimentExecution.State,
@@ -77,6 +91,108 @@ func onExperimentCompleted(w http.ResponseWriter, r *http.Request, body []byte) 
 
 	SendDatadogEvent(r.Context(), &config.Config, datadogEventBody)
 
+	exthttp.WriteBody(w, "{}")
+}
+
+var lastStartedSteps = make(map[float32]time.Time)
+var lastStartedStepsMux = &sync.RWMutex{}
+var lastCompletedSteps = make(map[float32]time.Time)
+var lastCompletedStepsMux = &sync.RWMutex{}
+
+func onExperimentStepStarted(w http.ResponseWriter, r *http.Request, body []byte) {
+	event, err := parseBodyToEventRequestBody(body)
+	if err != nil {
+		exthttp.WriteError(w, extension_kit.ToError("Failed to decode event request body", err))
+		return
+	}
+	lastStartedStepsMux.RLock()
+	lastStartedStep, ok := lastStartedSteps[event.ExperimentExecution.ExecutionId]
+	lastStartedStepsMux.RUnlock()
+	if !ok {
+		lastStartedStep = time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC)
+	}
+
+	for _, step := range event.ExperimentExecution.Steps {
+		if step.Type != event_kit_api.Wait && step.StartedTime != nil && step.StartedTime.After(lastStartedStep) {
+			lastStartedStepsMux.Lock()
+			lastStartedSteps[event.ExperimentExecution.ExecutionId] = *step.StartedTime
+			lastStartedStepsMux.Unlock()
+			tags := convertSteadybitEventToDataDogEventTags(event)
+			if tags == nil {
+				return
+			}
+			tags = append(tags, getStepTags(step)...)
+			var targetSummary string
+			for _, target := range *step.TargetExecutions {
+				targetSummary += fmt.Sprintf("\n- `%s`", target.TargetName)
+			}
+
+			datadogEventBody := datadogV1.EventCreateRequest{
+				Title: fmt.Sprintf("Experiment '%s' - Step started", event.ExperimentExecution.ExperimentKey),
+				Text: fmt.Sprintf("%%%%%% \nExperiment `%s` - `%s` (execution `%.0f`) - Step `%s` started.\n\nTargets:%s\n\n_The experiment is executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event)._\n %%%%%%",
+					event.ExperimentExecution.ExperimentKey,
+					event.ExperimentExecution.Name,
+					event.ExperimentExecution.ExecutionId,
+					*step.ActionId,
+					targetSummary,
+				),
+				Tags:           tags,
+				SourceTypeName: extutil.Ptr("Steadybit"),
+			}
+
+			SendDatadogEvent(r.Context(), &config.Config, datadogEventBody)
+		}
+	}
+	exthttp.WriteBody(w, "{}")
+}
+
+func onExperimentStepCompleted(w http.ResponseWriter, r *http.Request, body []byte) {
+	event, err := parseBodyToEventRequestBody(body)
+	if err != nil {
+		exthttp.WriteError(w, extension_kit.ToError("Failed to decode event request body", err))
+		return
+	}
+	lastCompletedStepsMux.RLock()
+	lastCompletedStep, ok := lastCompletedSteps[event.ExperimentExecution.ExecutionId]
+	lastCompletedStepsMux.RUnlock()
+	if !ok {
+		lastCompletedStep = time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC)
+	}
+
+	for _, step := range event.ExperimentExecution.Steps {
+		if step.Type != event_kit_api.Wait && step.EndedTime != nil && step.EndedTime.After(lastCompletedStep) && step.StartedTime != nil {
+			lastCompletedStepsMux.Lock()
+			lastCompletedSteps[event.ExperimentExecution.ExecutionId] = *step.EndedTime
+			lastCompletedStepsMux.Unlock()
+			tags := convertSteadybitEventToDataDogEventTags(event)
+			if tags == nil {
+				return
+			}
+			tags = append(tags, getStepTags(step)...)
+			duration := step.EndedTime.Sub(*step.StartedTime)
+			var targetSummary string
+			for _, target := range *step.TargetExecutions {
+				targetSummary += fmt.Sprintf("\n- `%s`", target.TargetName)
+			}
+
+			datadogEventBody := datadogV1.EventCreateRequest{
+				Title: fmt.Sprintf("Experiment '%s' - Step ended", event.ExperimentExecution.ExperimentKey),
+				Text: fmt.Sprintf("%%%%%% \nExperiment `%s` - `%s` (execution `%.0f`) - Step `%s` ended with state `%s` after %.2f seconds.\n\nTargets:%s\n\n_The experiment is executed through [Steadybit](https://steadybit.com/?utm_campaign=extension-datadog&utm_source=extension-datadog-event)._\n %%%%%%",
+					event.ExperimentExecution.ExperimentKey,
+					event.ExperimentExecution.Name,
+					event.ExperimentExecution.ExecutionId,
+					*step.ActionId,
+					step.State,
+					duration.Seconds(),
+					targetSummary,
+				),
+				Tags:           tags,
+				SourceTypeName: extutil.Ptr("Steadybit"),
+			}
+
+			SendDatadogEvent(r.Context(), &config.Config, datadogEventBody)
+		}
+	}
 	exthttp.WriteBody(w, "{}")
 }
 
@@ -91,7 +207,7 @@ func convertSteadybitEventToDataDogEventTags(event event_kit_api.EventRequestBod
 		"team_key:" + event.Team.Key,
 		"tenant_name:" + event.Tenant.Name,
 		"tenant_key:" + event.Tenant.Key,
-		"execution_id:" + fmt.Sprintf("%f", event.ExperimentExecution.ExecutionId),
+		"execution_id:" + fmt.Sprintf("%g", event.ExperimentExecution.ExecutionId),
 		"experiment_key:" + event.ExperimentExecution.ExperimentKey,
 		"experiment_name:" + event.ExperimentExecution.Name,
 		string("execution_state:" + event.ExperimentExecution.State),
@@ -130,6 +246,26 @@ func convertSteadybitEventToDataDogEventTags(event event_kit_api.EventRequestBod
 		tags = append(tags, "ended_time:"+event.ExperimentExecution.EndedTime.Format(time.RFC3339))
 	}
 
+	return tags
+}
+
+func getStepTags(step event_kit_api.ExperimentStepExecution) []string {
+	var tags []string
+	tags = append(tags, "step_type:"+string(step.Type))
+	tags = append(tags, "step_state:"+string(step.State))
+	if step.StartedTime != nil {
+		tags = append(tags, "step_started_time:"+step.StartedTime.Format(time.RFC3339))
+	}
+	if step.EndedTime != nil {
+		tags = append(tags, "step_ended_time:"+step.EndedTime.Format(time.RFC3339))
+	}
+	if step.Type == event_kit_api.Action {
+		tags = append(tags, "step_action_id:"+*step.ActionId)
+		tags = append(tags, "step_target_count:"+strconv.Itoa(len(*step.TargetExecutions)))
+		for targetIndex, target := range *step.TargetExecutions {
+			tags = append(tags, fmt.Sprintf("step_target_%d_name:%s", targetIndex, target.TargetName))
+		}
+	}
 	return tags
 }