+++ main.go
@@ -88,6 +88,18 @@ func getExtensionList() ExtensionListResponse {
 					ListenTo:   []string{"experiment.execution.completed", "experiment.execution.failed", "experiment.execution.canceled", "experiment.execution.errored"},
 					RestrictTo: extutil.Ptr(event_kit_api.Leader),
 				},
+				{
+					Method:     "POST",
+					Path:       "/events/experiment-step-started",
+					ListenTo:   []string{"experiment.execution.step-started"},
+					RestrictTo: extutil.Ptr(event_kit_api.Leader),
+				},
+				{
+					Method:     "POST",
+					Path:       "/events/experiment-step-completed",
+					ListenTo:   []string{"experiment.execution.step-completed", "experiment.execution.step-canceled", "experiment.execution.step-errored", "experiment.execution.step-failed"},
+					RestrictTo: extutil.Ptr(event_kit_api.Leader),
+				},
 			},
 		},
 	}